Improve support for mirroring the full keyspace
This makes a couple of minor changes to Etcd2Writer to ensure that the
transformed (replicated) keys are always well formed, even when one or
both of the src and dst prefix are the full keyspace (/).

Technically, etcdmirror works for this use case without modification,
but this is due to etcd's use of redirects (301) to canonicalize key
paths. To avoid the extra round-trip, we need to avoid the superfluous
slashes.

Note: Replication index key format is unchanged in the case of partial
keyspace mirroring (i.e., what we have now). In the full-keyspace case,
we use `/__replication/__ROOT__` rather than `/__replication`, to ensure
that the latter remains a dir.

This avoids manual cleanup actions when switching to / from full
keyspace mirroring.

Bug: T358636
Change-Id: If796d4f9f30fd8e08db2cda42fc524fd23a59370