Add support for an optional ignored-keys pattern
The --src-ignore-keys-regex flag can now be used to suppress replication
events for parts of the keyspace. Even when an event is suppressed, the
replication index still advances.

This is useful when replicating a larger prefix in the presence of
subpaths that are not meaningful when replicated (e.g., in-order keys
used by python-etcd locks).

Bug: T358636
Change-Id: I1732a020f8568b382cbf100d7d1b219f2a86cc1f